requirements.txt
PySocks>=1.7.1
requests
beautifulsoup4
dulwich
requests-pkcs12